pyproject.toml
[project]
name = "clube-dos-tampinha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "cassandra-driver>=3.29.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "st-pages>=1.0.1",
    "streamlit>=1.41.1",
]
